pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whatsnext"
version = "0.0.1"
description = "A job queue and task management system with client-server architecture"
authors = [{ name = "Cornelius Emde", email = "ai@corneliusemde.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
# Core has no dependencies - client and server are optional
dependencies = []

[project.optional-dependencies]
# Server dependencies (FastAPI + PostgreSQL)
server = [
    "fastapi>=0.100.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.20.0",
]

# Client dependencies
client = [
    "requests>=2.28.0",
    "tabulate>=0.9.0",
]

# CLI dependencies
cli = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "whatsnext[client]",
]

# All dependencies
all = ["whatsnext[server,client,cli]"]

[project.scripts]
whatsnext = "whatsnext.cli:main"
wnxt = "whatsnext.cli:main"

[project.urls]
Homepage = "https://github.com/cemde/WhatsNext"
Issues = "https://github.com/cemde/WhatsNext/issues"
Documentation = "https://whatsnext.readthedocs.io/"

[dependency-groups]
# Development tools (linting, testing, type checking)
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "httpx>=0.27.0",
    "ruff>=0.14.0",
    "ty>=0.0.5",
]

# Documentation building
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
    "pymdown-extensions>=10.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["whatsnext*"]

[tool.ruff]
line-length = 144
exclude = []

[tool.ruff.lint]
select = ["E", "F", "I"]  # E/F: pyflakes/pycodestyle, I: isort

[tool.ruff.lint.isort]
known-first-party = ["whatsnext"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "server: Tests that require server dependencies",
    "client: Tests that require client dependencies",
    "integration: Integration tests requiring both server and client",
]

[tool.coverage.run]
relative_files = true
source = ["whatsnext"]
omit = ["*/tests/*", "*/__pycache__/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2

[tool.ty.src]
exclude = []
